Friday Evening (8/11)
From 7PM CST to 10:37PM CST.

Built the tile system, got the player camera to work, got the player
character to be able to walk around within the confines of the screen
resolution (with an outline of the screen left untouched, will be used
for in game GUI's later), made a simple test map to walk around in
(255x255 in size).

Let's just hope that this is a solid foundation to work off of for the
next week, and not something I'll have to rip out tomorrow.