Saturday (8/12)
From (roughly) 6PM CST to 10:20PM CST.

Built the core of the map editor, extended the size of the hexadecimal
entries in the .rpmap files from 8 to 10 to accommodate the 2 digit
metadata property, did misc work on the camera system, added the ability
to save a map, as well as multiple misc changes.

So far so good. Tomorrow I'm going to try and get fog of war to work,
then if I have the time I'll see if I can start filling out the
tile/entity list.