Tuesday (8/15)
Worked a little on it throughout the day, no set time.

Multiple bug fixes, added the A* API, made the first entity (a rat),
started work on the faction system, started work on getting entities to
detect when they see the player.

Hopefully this foundation will be solid enough to build the combat
system off of. I should have combat done either tomorrow or the day
after.