Wednesday (8/16)
Worked on it through out the day.

Changed the direction of the game to something I should be able to
finish tomorrow. Added in gates, keys and pressure pads (as well as
rocks to put on pressure pads). Added a delete button to the mapmaker to
remove tiles and entities from a map. Made a few bug fixes and misc
changes.

If all goes as planned, I should be able to finish the game on Thursday,
then polish on Friday.